# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DevLens API
  version: 1.0.0
  description: |
    API contract สำหรับ DevLens Engineering Intelligence Platform

    Related documents:
    - 01-business-logic-architecture.md
    - 03-backend-roadmap.md
    - 04-metric-definitions.md
    - 05-database-design.dbml
    - 06-api-design.md
servers:
  - url: /api/v1
    description: Current API version
...
security:
  - bearerAuth: []
paths:
...
  /repositories/{repositoryId}/sync:
    post:
      tags:
        - Sync Jobs
      summary: Start repository synchronization
      operationId: createRepositorySync
      parameters:
        - $ref: "#/components/parameters/RepositoryId"
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSyncRequest"
      responses:
        "202":
          description: Sync job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncJobResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /repositories/{repositoryId}/sync-jobs:
    get:
      tags:
        - Sync Jobs
      summary: Get repository synchronization history
      operationId: listRepositorySyncJobs
      parameters:
        - $ref: "#/components/parameters/RepositoryId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/SyncStatus"
        - $ref: "#/components/parameters/SortOrder"
      responses:
        "200":
          description: Sync job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncJobListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /sync-jobs/{syncJobId}:
    get:
      tags:
        - Sync Jobs
      summary: Get synchronization job detail
      operationId: getSyncJob
      parameters:
        - $ref: "#/components/parameters/SyncJobId"
      responses:
        "200":
          description: Sync job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncJobResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /sync-jobs/{syncJobId}/retry:
    post:
      tags:
        - Sync Jobs
      summary: Retry a failed synchronization job
      operationId: retrySyncJob
      parameters:
        - $ref: "#/components/parameters/SyncJobId"
      responses:
        "202":
          description: Sync job re-queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncJobResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /sync-jobs/{syncJobId}/cancel:
    post:
      tags:
        - Sync Jobs
      summary: Cancel a pending or running synchronization job
      operationId: cancelSyncJob
      parameters:
        - $ref: "#/components/parameters/SyncJobId"
      responses:
        "202":
          description: Sync job canceled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncJobResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
...
    RepositoryId:
      name: repositoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SyncJobId:
      name: syncJobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    SortOrder:
      name: sortOrder
      in: query
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
...
    SyncStatus:
      name: status
      in: query
      schema:
        type: string
        enum:
          - pending
          - running
          - completed
          - failed
...
  responses:
    Unauthorized:
      description: Authentication is required or token is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: User does not have permission to access the resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Request conflicts with current resource state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    PaginationMeta:
      type: object
      required:
        - page
        - pageSize
        - totalItems
        - totalPages
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3
...
    CreateSyncRequest:
      type: object
      properties:
        mode:
          type: string
          enum:
            - incremental
            - full
          default: incremental
        from:
          type: string
          format: date
          nullable: true
    SyncJob:
      type: object
      required:
        - id
        - repositoryId
        - status
        - progress
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        repositoryId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - pending
            - running
            - completed
            - failed
            - canceled
        progress:
          type: integer
          minimum: 0
          maximum: 100
        triggeredBy:
          type: string
          format: uuid
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
    SyncJobResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/SyncJob"
    SyncJobListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SyncJob"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: REPOSITORY_NOT_FOUND
        message:
          type: string
          example: Repository not found
        requestId:
          type: string
          example: req_01JABCXYZ
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
...